requirements.txt
fastapi
uvicorn
python-multipart
faiss-cpu
numpy
openai
python-dotenv
langchain_community
langchain